+++ src/Conversions/Conversion.php
@@ -5,13 +5,14 @@ namespace Spatie\MediaLibrary\Conversions;
 use BadMethodCallException;
 use Spatie\Image\Manipulations;
 use Spatie\MediaLibrary\MediaCollections\Models\Media;
+use Spatie\MediaLibrary\Support\FileNamer\FileNamer;
 
 /** @mixin \Spatie\Image\Manipulations */
 class Conversion
 {
     protected string $name = '';
 
-    protected ConversionFileNamer $conversionFileNamer;
+    protected FileNamer $fileNamer;
 
     protected float $extractVideoFrameAtSecond = 0;
 
@@ -19,7 +20,7 @@ class Conversion
 
     protected array $performOnCollections = [];
 
-    protected bool $performOnQueue = true;
+    protected bool $performOnQueue;
 
     protected bool $keepOriginalImageFormat = false;
 
@@ -37,9 +38,11 @@ class Conversion
             ->optimize(config('media-library.image_optimizers'))
             ->format(Manipulations::FORMAT_JPG);
 
-        $this->conversionFileNamer = app(config('media-library.conversion_file_namer'));
+        $this->fileNamer = app(config('media-library.file_namer'));
 
         $this->loadingAttributeValue = config('media-library.default_loading_attribute_value');
+
+        $this->performOnQueue = config('media-library.queue_conversions_by_default', true);
     }
 
     public static function create(string $name)
@@ -222,8 +225,10 @@ class Conversion
 
     public function getConversionFile(Media $media): string
     {
-        $fileName = $this->conversionFileNamer->getFileName($this, $media);
-        $extension = $this->conversionFileNamer->getExtension($this, $media);
+        $fileName = $this->fileNamer->conversionFileName($media->file_name, $this);
+
+        $fileExtension = $this->fileNamer->extensionFromBaseImage($media->file_name);
+        $extension = $this->getResultExtension($fileExtension) ?: $fileExtension;
 
         return "{$fileName}.{$extension}";
     }