+++ src/MediaCollections/Models/Media.php
@@ -5,7 +5,9 @@ namespace Spatie\MediaLibrary\MediaCollections\Models;
 use DateTimeInterface;
 use Illuminate\Contracts\Support\Htmlable;
 use Illuminate\Contracts\Support\Responsable;
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\MorphTo;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
@@ -21,8 +23,10 @@ use Spatie\MediaLibrary\MediaCollections\Models\Concerns\HasUuid;
 use Spatie\MediaLibrary\MediaCollections\Models\Concerns\IsSorted;
 use Spatie\MediaLibrary\ResponsiveImages\RegisteredResponsiveImages;
 use Spatie\MediaLibrary\Support\File;
+use Spatie\MediaLibrary\Support\MediaLibraryPro;
 use Spatie\MediaLibrary\Support\TemporaryDirectory;
 use Spatie\MediaLibrary\Support\UrlGenerator\UrlGeneratorFactory;
+use Spatie\MediaLibraryPro\Models\TemporaryUpload;
 
 class Media extends Model implements Responsable, Htmlable
 {
@@ -39,6 +43,7 @@ class Media extends Model implements Responsable, Htmlable
     protected $casts = [
         'manipulations' => 'array',
         'custom_properties' => 'array',
+        'generated_conversions' => 'array',
         'responsive_images' => 'array',
     ];
 
@@ -182,6 +187,38 @@ class Media extends Model implements Responsable, Htmlable
         return $conversions->map(fn (Conversion $conversion) => $conversion->getName())->toArray();
     }
 
+    public function getGeneratedConversions(): Collection
+    {
+        return collect($this->generated_conversions ?? []);
+    }
+
+
+    public function markAsConversionGenerated(string $conversionName): self
+    {
+        $generatedConversions = $this->generated_conversions;
+
+        Arr::set($generatedConversions, $conversionName, true);
+
+        $this->generated_conversions = $generatedConversions;
+
+        $this->save();
+
+        return $this;
+    }
+
+    public function markAsConversionNotGenerated(string $conversionName): self
+    {
+        $generatedConversions = $this->generated_conversions;
+
+        Arr::set($generatedConversions, $conversionName, false);
+
+        $this->generated_conversions = $generatedConversions;
+
+        $this->save();
+
+        return $this;
+    }
+
     public function hasGeneratedConversion(string $conversionName): bool
     {
         $generatedConversions = $this->getGeneratedConversions();
@@ -189,19 +226,6 @@ class Media extends Model implements Responsable, Htmlable
         return $generatedConversions[$conversionName] ?? false;
     }
 
-    public function markAsConversionGenerated(string $conversionName, bool $generated): self
-    {
-        $this->setCustomProperty("generated_conversions.{$conversionName}", $generated);
-
-        $this->save();
-
-        return $this;
-    }
-
-    public function getGeneratedConversions(): Collection
-    {
-        return collect($this->getCustomProperty('generated_conversions', []));
-    }
 
     public function toResponse($request)
     {
@@ -209,7 +233,7 @@ class Media extends Model implements Responsable, Htmlable
             'Cache-Control' => 'must-revalidate, post-check=0, pre-check=0',
             'Content-Type' => $this->mime_type,
             'Content-Length' => $this->size,
-            'Content-Disposition' => 'attachment; filename="'.$this->file_name.'"',
+            'Content-Disposition' => 'attachment; filename="' . $this->file_name . '"',
             'Pragma' => 'public',
         ];
 
@@ -252,7 +276,7 @@ class Media extends Model implements Responsable, Htmlable
     {
         $temporaryDirectory = TemporaryDirectory::create();
 
-        $temporaryFile = $temporaryDirectory->path('/').DIRECTORY_SEPARATOR.$this->file_name;
+        $temporaryFile = $temporaryDirectory->path('/') . DIRECTORY_SEPARATOR . $this->file_name;
 
         /** @var \Spatie\MediaLibrary\MediaCollections\Filesystem $filesystem */
         $filesystem = app(Filesystem::class);
@@ -262,6 +286,7 @@ class Media extends Model implements Responsable, Htmlable
         $newMedia = $model
             ->addMedia($temporaryFile)
             ->usingName($this->name)
+            ->setOrder($this->order_column)
             ->withCustomProperties($this->custom_properties)
             ->toMediaCollection($collectionName, $diskName);
 
@@ -298,5 +323,26 @@ class Media extends Model implements Responsable, Htmlable
     public function __invoke(...$arguments): HtmlableMedia
     {
         return $this->img(...$arguments);
+    }
+
+    public function temporaryUpload(): BelongsTo
+    {
+        MediaLibraryPro::ensureInstalled();
+
+        return $this->belongsTo(TemporaryUpload::class);
+    }
+
+    public static function findWithTemporaryUploadInCurrentSession(array $uuids)
+    {
+        MediaLibraryPro::ensureInstalled();
+
+        return static::query()
+            ->whereIn('uuid', $uuids)
+            ->whereHasMorph(
+                'model',
+                [TemporaryUpload::class],
+                fn (Builder $builder) => $builder->where('session_id', session()->getId())
+            )
+            ->get();
     }
 }