+++ src/Support/FileNamer/DefaultFileNamer.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Spatie\MediaLibrary\Support\FileNamer;
+
+use Spatie\MediaLibrary\Conversions\Conversion;
+
+class DefaultFileNamer extends FileNamer
+{
+    public function conversionFileName(string $fileName, Conversion $conversion): string
+    {
+        $strippedFileName = pathinfo($fileName, PATHINFO_FILENAME);
+
+        return "{$strippedFileName}-{$conversion->getName()}";
+    }
+
+    public function responsiveFileName(string $fileName): string
+    {
+        return pathinfo($fileName, PATHINFO_FILENAME);
+    }
+}